+++ apps/opplaringspenger-soknad/src/app/components/calendar-grid/calendarGridMessages.ts
@@ -1,13 +1,27 @@
+const nb = {
+    'calendarGrid.Mandag': 'Mandag',
+    'calendarGrid.Tirsdag': 'Tirsdag',
+    'calendarGrid.Onsdag': 'Onsdag',
+    'calendarGrid.Torsdag': 'Torsdag',
+    'calendarGrid.Fredag': 'Fredag',
+    'calendarGrid.uke': 'uke',
+    'calendarGrid.Uke': 'Uke',
+    'calendarGrid.måned': 'måned',
+    'calendarGrid.Måned': 'Måned',
+};
+
+const nn: Record<keyof typeof nb, string> = {
+    'calendarGrid.Mandag': 'Måndag',
+    'calendarGrid.Tirsdag': 'Tysdag',
+    'calendarGrid.Onsdag': 'Onsdag',
+    'calendarGrid.Torsdag': 'Torsdag',
+    'calendarGrid.Fredag': 'Fredag',
+    'calendarGrid.uke': 'veke',
+    'calendarGrid.Uke': 'Veke',
+    'calendarGrid.måned': 'månad',
+    'calendarGrid.Måned': 'Månad',
+};
 export const calendarGridMessages = {
-    nb: {
-        'calendarGrid.Mandag': 'Mandag',
-        'calendarGrid.Tirsdag': 'Tirsdag',
-        'calendarGrid.Onsdag': 'Onsdag',
-        'calendarGrid.Torsdag': 'Torsdag',
-        'calendarGrid.Fredag': 'Fredag',
-        'calendarGrid.uke': 'uke',
-        'calendarGrid.Uke': 'Uke',
-        'calendarGrid.måned': 'måned',
-        'calendarGrid.Måned': 'Måned',
-    },
+    nb,
+    nn,
 };