+++ apps/opplaringspenger-soknad/src/app/pages/kvittering/kvitteringMesssages.ts
@@ -9,6 +9,15 @@ const nb = {
         'Når søknaden er ferdig behandlet får du et svar fra oss på <MinSideLenke>Min side</MinSideLenke>. Du kan sjekke <SaksbehandlingstidLenke>saksbehandlingstiden her</SaksbehandlingstidLenke>.',
 };
 
-const nn: Record<keyof typeof nb, string> = { ...nb };
+const nn: Record<keyof typeof nb, string> = {
+    'page.kvittering.sidetittel': 'Vi har motteke søknad om opplæringspengar',
+    'page.kvittering.tittel': 'Vi har motteke søknad om opplæringspengar',
+    'page.kvittering.info.tittel': 'Kva skjer vidare no?',
+    'page.kvittering.list.item.1': 'Du må be arbeidsgivar om å sende inntektsmelding',
+    'page.kvittering.list.item.2':
+        'Vi startar behandlinga av søknaden din når vi har motteke all naudsynt dokumentasjon. Vi kontaktar deg dersom vi treng fleire opplysningar.',
+    'page.kvittering.list.item.3':
+        'Når søknaden er ferdig behandla får du eit svar frå oss på <MinSideLenke>Min side</MinSideLenke>. Du kan sjekke <SaksbehandlingstidLenke>saksbehandlingstida her</SaksbehandlingstidLenke>.',
+};
 
 export const kvitteringMessages = { nb, nn };