+++ apps/opplaringspenger-soknad/src/app/pages/velkommen/velkommenPageMessages.ts
@@ -3,7 +3,8 @@ const nb = {
     'page.velkommen.guide.tittel': 'Hei {navn}',
     'page.velkommen.guide.ingress': 'Velkommen til søknad om opplæringspenger.',
     'page.velkommen.guide.tekst.1':
-        'Denne søknaden er for deg som må være borte fra jobb for å ta opplæring som er nødvendig for å kunne ta deg av et barn med langvarig sykdom eller funksjonshemning.',
+        'Denne søknaden er for deg som må være borte fra jobb på grunn av opplæring som er nødvendig for å kunne ta deg av et barn med langvarig sykdom eller funksjonshemning.',
+    'page.velkommen.guide.tekst.2': 'Du kan lese <Lenke>mer om opplæringspenger her</Lenke>.',
 
     'page.velkommen.omSøknaden.tittel': 'Om søknaden',
     'page.velkommen.omSøknaden.1': 'Du får veiledning underveis i søknaden om hva du skal fylle ut, og hvordan.',
@@ -23,20 +24,56 @@ const nb = {
     'personopplysninger.2':
         'Vi innhenter og mottar opplysninger om deg for å behandle saken din. Det er nødvendig for at du skal få riktig tjeneste.',
     'personopplysninger.3': 'Hvilke opplysninger innhenter vi?',
-    'personopplysninger.4': 'Opplysningene vi innhenter kommer enten fra deg eller fra offentlige registre:',
+    'personopplysninger.4':
+        'Opplysningene vi innhenter kommer enten fra deg, arbeidsgiver eller fra offentlige registre:',
     'personopplysninger.4.1': 'opplysninger om helsen til den du skal søke opplæringspenger for',
     'personopplysninger.4.2': 'arbeidsforholdene dine og inntekten din',
     'personopplysninger.4.3': 'ytelser du mottar fra Nav',
     'personopplysninger.4.4': 'tilknytningen din til Norge',
     'personopplysninger.4.5':
         'trygdeordninger du kan ha rett til i andre land. Vi kan også sende opplysninger om deg til trygdemyndigheter i andre land',
-    'personopplysninger.5.1':
-        'Vi lagrer opplysninger vi får inn om personen du søker opplæringspenger for. Dette gjør vi for å kunne gjenbruke opplysningene hvis andre søker opplæringspenger for samme person.',
 
     'personopplysninger.6':
         'Du har rett til innsyn i saken din. Vil du vite mer om hvordan Nav behandler personopplysninger? Se <Lenke>nav.no/personvern</Lenke>.',
 };
 
-const nn: Record<keyof typeof nb, string> = { ...nb };
+const nn: Record<keyof typeof nb, string> = {
+    'page.velkommen.sidetittel': 'Søknad om opplæringspengar',
+    'page.velkommen.guide.tittel': 'Hei {navn}',
+    'page.velkommen.guide.ingress': 'Velkomen til søknad om opplæringspengar.',
+    'page.velkommen.guide.tekst.1':
+        'Denne søknaden er for deg som må vere borte frå jobb på grunn av opplæring som er nødvendig for å kunne ta deg av eit barn med langvarig sjukdom eller funksjonshemming.',
+    'page.velkommen.guide.tekst.2': 'Du kan lese <Lenke>meir om opplæringspengar her</Lenke>.',
 
+    'page.velkommen.omSøknaden.tittel': 'Om søknaden',
+    'page.velkommen.omSøknaden.1': 'Du får rettleiing undervegs i søknaden om kva du skal fylle ut, og korleis.',
+    'page.velkommen.omSøknaden.2':
+        'Me tek vare på svara dine i 72 timar. Så, om du innanfor den tida til dømes vil ta ein pause eller blir automatisk logga ut, fortset du der du var når du kjem tilbake.',
+    'page.velkommen.omSøknaden.3':
+        'Du må svare på alle spørsmåla for å kunne gå vidare. Om du manglar etterspurd dokumentasjon, kan du ettersende det så snart du kan.',
+    'page.velkommen.omSøknaden.4': 'Om korleis me hentar opplysningar om deg',
+    'page.velkommen.form.bekreftLabel': 'Eg bekreftar at eg har forstått mitt ansvar som søkjar',
+    'page.velkommen.form.ansvar.tittel': 'Ditt ansvar som søkjar',
+    'page.velkommen.form.ansvar.list.1':
+        'Eg forstår at om eg gir uriktige opplysningar, kan det få konsekvensar for retten min til det eg søkjer om.',
+    'page.velkommen.form.ansvar.list.2.1': 'Eg har lese og forstått det som står på',
+    'page.velkommen.form.ansvar.list.2.2': 'nav.no/rett og plikt',
+    'page.velkommen.harForståttRettigheterOgPlikter.notChecked':
+        'Du må velje at du har forstått ditt ansvar som søkjar',
+    'personopplysninger.1': 'Slik behandlar Nav personopplysningane dine',
+    'personopplysninger.2':
+        'Me hentar inn og mottar opplysningar om deg for å behandle saka di. Det er nødvendig for at du skal få riktig teneste.',
+    'personopplysninger.3': 'Kva for opplysningar hentar me inn?',
+    'personopplysninger.4':
+        'Opplysningane me hentar inn kjem anten frå deg, arbeidsgivar eller frå offentlege register:',
+    'personopplysninger.4.1': 'opplysningar om helsa til den du skal søkje opplæringspengar for',
+    'personopplysninger.4.2': 'arbeidsforholda dine og inntekta di',
+    'personopplysninger.4.3': 'ytingar du mottar frå Nav',
+    'personopplysninger.4.4': 'tilknytinga di til Noreg',
+    'personopplysninger.4.5':
+        'trygdeordningar du kan ha rett til i andre land. Me kan også sende opplysningar om deg til trygdemyndigheiter i andre land.',
+
+    'personopplysninger.6':
+        'Du har rett til innsyn i saka di. Vil du vite meir om korleis Nav behandlar personopplysningar? Sjå <Lenke>nav.no/personvern</Lenke>.',
+};
 export const velkommenPageMessages = { nb, nn };