+++ apps/opplaringspenger-soknad/src/app/søknad/steps/oppsummering/arbeid-i-perioden-summary/ArbeidIPeriodenSummaryItem.tsx
@@ -17,7 +17,6 @@ interface Props {
     valgteDatoer: Date[];
     arbeidIPeriode: ArbeidIPeriodeApiData;
     normaltimerUke: number;
-    harFlereArbeidsforhold: boolean;
 }
 
 export interface ArbeidIPeriodenSummaryItemType extends ArbeidsforholdApiData {
@@ -31,28 +30,23 @@ const fjernDagerIkkeSøktForOgUtenArbeidstid = (enkeltdager: TidEnkeltdagApiData
     });
 };
 
-const ArbeidIPeriodeSummaryItem: React.FC<Props> = ({ arbeidIPeriode, valgteDatoer, harFlereArbeidsforhold }) => {
+const ArbeidIPeriodeSummaryItem: React.FC<Props> = ({ arbeidIPeriode, valgteDatoer }) => {
     return (
         <>
-            {harFlereArbeidsforhold && (
-                <>
-                    {(arbeidIPeriode.jobberIPerioden === JobberIPeriodeSvar.heltFravær ||
-                        arbeidIPeriode.jobberIPerioden === JobberIPeriodeSvar.somVanlig) && (
-                        <p style={{ marginTop: 0 }}>
-                            <AppText
-                                id={`oppsummering.arbeidIPeriode.jobberIPerioden.${arbeidIPeriode.jobberIPerioden}`}
-                            />
-                        </p>
-                    )}
-                    {arbeidIPeriode.jobberIPerioden === JobberIPeriodeSvar.redusert && (
-                        <p style={{ marginTop: 0 }}>
-                            <AppText
-                                id={`oppsummering.arbeidIPeriode.jobberIPerioden.${arbeidIPeriode.jobberIPerioden}`}
-                            />
-                        </p>
-                    )}
-                </>
-            )}
+            <>
+                {(arbeidIPeriode.jobberIPerioden === JobberIPeriodeSvar.heltFravær ||
+                    arbeidIPeriode.jobberIPerioden === JobberIPeriodeSvar.somVanlig) && (
+                    <p style={{ marginTop: 0 }}>
+                        <AppText id={`oppsummering.arbeidIPeriode.jobberIPerioden.${arbeidIPeriode.jobberIPerioden}`} />
+                    </p>
+                )}
+                {arbeidIPeriode.jobberIPerioden === JobberIPeriodeSvar.redusert && (
+                    <p style={{ marginTop: 0 }}>
+                        <AppText id={`oppsummering.arbeidIPeriode.jobberIPerioden.${arbeidIPeriode.jobberIPerioden}`} />
+                    </p>
+                )}
+            </>
+
             {arbeidIPeriode.jobberIPerioden === JobberIPeriodeSvar.redusert && arbeidIPeriode.enkeltdager && (
                 <Block margin="xl">
                     <Heading size="xsmall" level="4" spacing={true}>