+++ apps/opplaringspenger-soknad/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @navikt/opplaringspenger-soknad
 
+## 1.17.1
+
+### Patch Changes
+
+- Trigger deploy etter pakkeoppdateringer
+- Updated dependencies [f2a18cf]
+- Updated dependencies
+- Updated dependencies [f2a18cf]
+- Updated dependencies [f2a18cf]
+    - @navikt/sif-common-utils@3.49.0
+    - @navikt/sif-app-register@0.1.6
+    - @navikt/sif-common-amplitude@2.27.1
+    - @navikt/sif-common-core-ds@10.2.1
+    - @navikt/sif-common-env@0.1.3
+    - @navikt/sif-common-formik-ds@2.4.0
+    - @navikt/sif-common-forms-ds@19.0.0
+    - @navikt/sif-common-hooks@0.3.1
+    - @navikt/sif-common-sentry@0.28.1
+    - @navikt/sif-common-soknad-ds@24.0.0
+    - @navikt/sif-common-ui@0.9.2
+
 ## 1.17.0
 
 ### Minor Changes