+++ packages/sif-common-amplitude/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/sif-common-amplitude
 
+## 2.27.1
+
+### Patch Changes
+
+- Trigger deploy etter pakkeoppdateringer
+
 ## 2.27.0
 
 ### Minor Changes